update composer lock/json to use product-manager 0.1.3 for fixes
Pre 0.1.2/0.1.3, the product-manager uses a static configured domain to
access the zed API backend. So we need the fix for a custom env.

As 0.1.3 is currently not marked as "stable", the composer constrains need
to be adjusted to "-dev". See https://getcomposer.org/doc/articles/versions.md#stability
for details.