Switch to PHP 7.2.5, NodeJS 10 and remove mcrypt and shop patches
Starting with PHP 7.2.x mcrypt is no longer a core extension and not
required by spryker anymore.
The claranet/php 7.2.x image uses Debian Stretch instead of Jessie,
so we can support NodeJS 10 as well.
Due to the 2.29 rabbitmq support, we don't need the patch which
prohibits the rabbitmq log handler. The patch also doesn't work with
spryker 2.30, so drop it.