Change secrets handling of local docker compose stack
Instead of writing into a single git-ignored file, we instead read a
file line per line and if this line ends with an `=` a password will
generated and appended to this line. The result is being written to a
hidden git-ignored dotfile. This approach makes it more transparent to
the user whats will be generated.

- Reimplemented pwgen in shell script
- Adds template slurp in and auto generation of effective secrets