Fix spryker upstream bugs in jenkins slave
Spryker are developing their stack with traditional, static servers in
mind instead of immutable containers. Thats why they are paying quite
some attention in trying not to conflict different environments with
each other. Due to the isolation of the containers this is not necessary
anymore. This commit tries to keep backward compatibility.