Extend .dockerignore to exclude sesible temporary secrets from image build
This security issue has low impact, but you should update to a version
which includes this commit!

If an docker image gets build AFTER the local stack was already run on
the system that builds the image, it might be, that the docker build
process adds the `docker/.secret.env` file which contains secrets for
a local build, dynamically created for the users local dev stack.