Fix Jenkins docker image build via docker-compose and revert split
This reverts the split of the Dockerfile for the Jenkins slave image
build.

Using the same Dockerfile, instead of letting the Jenkins slave image depend on
the main shop image, allows to paralelize the build process, if required; it helps
a lot in different CI systems (e.g. cloud builder on GCP) and it makes maintaining
the Jenkins Slave image no different to the spryker container stack.

Last but not least it makes it easier to get an overview on which stage jenkins components need
to be added and this is done alltogether.

In addition it fixes the broken jenkins docker build when building via docker-compose.